Use kwargs in function calls
This lets us thread keywords through multiple functions without needing
to explicitly pass them to the next function. It will make a which()
function easy to implement in the future because we can just use number
between function for the range of indices and pass the chocies through
to ask to have their enumeration pretty printed.